change postfix in tests class Tests -> Test
